// src/enums/csosn-cf.enum.ts
export enum CsosnCfEnum {
  CSOSN_CSOSNCF_101 = "101",
  CSOSN_CSOSNCF_102 = "102",
  CSOSN_CSOSNCF_103 = "103",
  CSOSN_CSOSNCF_201 = "201",
  CSOSN_CSOSNCF_202 = "202",
  CSOSN_CSOSNCF_203 = "203",
  CSOSN_CSOSNCF_300 = "300",
  CSOSN_CSOSNCF_400 = "400",
  CSOSN_CSOSNCF_500 = "500",
  CSOSN_CSOSNCF_900 = "900",
}